requirements.txt
# F1 Race Knowledge Base & Vector Database
# Python 3.9+

# Data Collection
fastf1>=3.3.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Vector Database
pinecone>=5.0.0

# Embeddings
openai>=1.12.0

# Data Processing & Analysis
scikit-learn>=1.3.0
scipy>=1.11.0

# Machine Learning (for predictions)
xgboost>=2.0.0
lightgbm>=4.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0

# Data Serialization
pyarrow>=15.0.0
orjson>=3.9.0

# API/Web (optional for future use)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Development
pytest>=8.0.0
black>=24.1.0
